fix: resolve mypy errors in run_report and temporal test
- Fix no-any-return in _span_duration_ms by wrapping result in float()
- Fix arg-type in _make_temporal_settings by annotating defaults dict
  as dict[str, object] to match **overrides type
- Install git pre-commit hook (scripts/pre-commit) which mirrors CI:
  ruff check, ruff format --check, mypy, unit tests with 85% coverage

+++ src/job_hunter_agents/observability/run_report.py
@@ -257,5 +257,5 @@ def _span_duration_ms(span: Any) -> float:  # noqa: ANN401
     start = getattr(span, "start_time", None)
     end = getattr(span, "end_time", None)
     if start is not None and end is not None:
-        return (end - start) / 1_000_000  # ns -> ms
+        return float((end - start) / 1_000_000)  # ns -> ms
     return 0.0
